requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# Configuration
pydantic-settings>=2.1.0

# OCR
paddleocr>=2.7.0
paddlepaddle>=2.5.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.1
Pillow>=10.2.0
